Cleaned up tests; added `parametrize` for reasoning tests to test opeani/anthropic
